refactor(codegen): extract compile_lvalue into lvalues.rs (#113 phase 5)
